fix aaaa

+++ algo/maxLikMPI.py
@@ -267,9 +267,9 @@ if __name__ == '__main__':
         if len(sys.argv)>1:
             dbname=sys.argv[1]
         if len(sys.argv)>2:
-            n_states=float(sys.argv[2])
+            n_states=int(sys.argv[2])
         if len(sys.argv)>3:
-            Sth=int(sys.argv[3])
+            Sth=float(sys.argv[3])
     dbname=comm.bcast(dbname,root=0)
     n_states=comm.bcast(n_states,root=0)
     Sth=comm.bcast(Sth,root=0)
